+++ __tests__/utils/loggers.spec.ts
@@ -1,5 +1,5 @@
 import LogLevel from '@microsoft/omnichannel-ic3core/lib/logging/LogLevel';
-import {ACSAdapterLogger, ACSClientLogger, IC3ClientLogger, OCSDKLogger} from '../../src/utils/loggers';
+import {ACSAdapterLogger, ACSClientLogger, IC3ClientLogger, OCSDKLogger, CallingSDKLogger} from '../../src/utils/loggers';
 
 describe('loggers', () => {
     describe('IC3ClientLogger', () => {
@@ -341,6 +341,73 @@ describe('loggers', () => {
             logger.completeScenario('');
 
             expect((logger as any).scenarioMarker.completeScenario).toHaveBeenCalledTimes(1);
+        });
+    });
+
+    describe('CallingSDKLogger', () => {
+        const omnichannelConfig = {
+            orgId: '',
+            orgUrl: '',
+            widgetId: ''
+        }
+
+        const telemetry = {
+            debug: jest.fn(),
+            warn: jest.fn(),
+            error: jest.fn(),
+            info: jest.fn()
+        }
+
+        it('CallingSDKLogger.logCallingSdkTelemetryEvent() with LogLevel DEBUG should call telemetry.debug()', () => {
+            const logger = new CallingSDKLogger(omnichannelConfig);
+
+            const eventData = {
+                Event: ''
+            };
+
+            logger.useTelemetry(telemetry as any);
+            logger.logCallingSdkTelemetryEvent(LogLevel.DEBUG, eventData as any);
+
+            expect(telemetry.debug).toBeCalledTimes(1);
+        });
+
+        it('CallingSDKLogger.logCallingSdkTelemetryEvent() with LogLevel WARN should call telemetry.warn()', () => {
+            const logger = new CallingSDKLogger(omnichannelConfig);
+
+            const eventData = {
+                Event: ''
+            };
+
+            logger.useTelemetry(telemetry as any);
+            logger.logCallingSdkTelemetryEvent(LogLevel.WARN, eventData as any);
+
+            expect(telemetry.warn).toBeCalledTimes(1);
+        });
+
+        it('CallingSDKLogger.logCallingSdkTelemetryEvent() with LogLevel ERROR should call telemetry.error()', () => {
+            const logger = new CallingSDKLogger(omnichannelConfig);
+
+            const eventData = {
+                Event: ''
+            };
+
+            logger.useTelemetry(telemetry as any);
+            logger.logCallingSdkTelemetryEvent(LogLevel.ERROR, eventData as any);
+
+            expect(telemetry.error).toBeCalledTimes(1);
+        });
+
+        it('CallingSDKLogger.logCallingSdkTelemetryEvent() with LogLevel INFO should call telemetry.info()', () => {
+            const logger = new CallingSDKLogger(omnichannelConfig);
+
+            const eventData = {
+                Event: ''
+            };
+
+            logger.useTelemetry(telemetry as any);
+            logger.logCallingSdkTelemetryEvent(LogLevel.INFO, eventData as any);
+
+            expect(telemetry.info).toBeCalledTimes(1);
         });
     });
 })