+++ src/core/ChatTranscriptBody.ts
@@ -1,5 +1,5 @@
 export default interface ChatTranscriptBody {
     emailAddress: string;
     attachmentMessage: string;
-    locale: string;
+    locale?: string;
 }