+++ CHANGELOG.md
@@ -13,6 +13,10 @@ All notable changes to this project will be documented in this file.
 - Add ability to automatically pass locale from chat config on calling `ChatSDK.emailLiveChatTranscript()`
 - Bubble up `WidgetUseOutsideOperatingHour` exception
 - Add `acs_webchat-chat-adapter` middleware to add default `channelData.tags` & `channelData.metadata`
+- Update `ChatConfig` interface with `msdyn_enablemarkdown` property
+- Throw exception on `ChatSDK.getVoiceVideoCalling()` if feature is disabled or platform is not supported
+- Add `participantType` & `canRenderPostChat` as response of getConversationDetails() API
+- Add support for separate bot post chat survey feature
 
 ### Fix
 - Add `acs_webchat-chat-adapter` middlewares to format `channelData.tags`
@@ -26,6 +30,7 @@ All notable changes to this project will be documented in this file.
 - README: added examples on usages of the post chat APIs.
 - Uptake [@azure/communication-chat@1.1.1](https://www.npmjs.com/package/@azure/communication-chat/v/1.1.1)
 - Uptake [acs_webchat-chat-adapter@0.0.35-beta.2](https://unpkg.com/acs_webchat-chat-adapter@0.0.35-beta.2/dist/chat-adapter.js)
+- Update `locale` property in `ChatTranscriptBody` interface to be optional
 
 ## [1.0.0] - 2021-10-08
 ### Added